Modify ci.sh to add new feature

+++ amazon/ci.sh
@@ -1,17 +1,66 @@
 #!/bin/bash
 
+function is_ignore()
+{
+  package=$1
+
+  rec=`grep "$package" excludes`
+  if [ -z "$rec" ]; then 
+    return 1
+  fi
+  return 0
+}
+
+function do_ci()
+{
+  package=$1
+  DIR=$2
+
+  rpm -q -l $package > $DIR/list
+  rpm -q -i $package > $DIR/info
+  rpm -q -R $package > $DIR/depend
+  rpm -q --changelog $package > $DIR/changelog
+  rpm -q --scripts $package > $DIR/scripts
+}
+
+function do_srpm()
+{
+  package=$1
+
+  ./scripts/getsrc.expect $package
+}
+
+# Set default Action
+ACTION="default"
+
+# Set Action from parameter
+if [ $# -gt 1 ]; then 
+  echo "./ci.sh [srpm]"
+  exit 0
+fi
+
+if [ $# -eq 0 ]; then
+  ACTION="default"
+elif [ $1 = "srpm" ]; then
+  ACTION="srpm"
+fi
+
+# Delete all Package information
+if [ $ACTION == "default" ]; then
+  rm -rf pkg/*
+elif [ $ACTION == "srpm" ]; then
+  rm -rf srpm/*
+fi 
+
 # Get All Package from system
 rpm -qa | sort > packages
 
-# Delete all Package information
-rm -rf pkg/*
-
 while read LINE
 do
-  # Ignore Some Exclude package
   package=${LINE%%-[0-9]*}
-  grep $package excludes 
-  if [ $? == 0 ]; then 
+
+  # Ignore Some Exclude package
+  if [ `is_ignore $package` ]; then 
     continue
   fi
 
@@ -24,10 +73,11 @@ do
   fi
   DIR=pkg/$package
 
-  # Collection infromation from each Package
-  rpm -q -l $package > $DIR/list
-  rpm -q -i $package > $DIR/info
-  rpm -q -R $package > $DIR/depend
-  rpm -q --changelog $package > $DIR/changelog
-  rpm -q --scripts $package > $DIR/scripts
+  if [ $ACTION == "default" ]; then
+    # Collection infromation from each Package
+    do_ci $package $DIR
+  elif [ $ACTION == "srpm" ]; then
+    # Down SRPM to /usr/src/srpm/debug/ 
+    do_srpm $package
+  fi 
 done < packages
